separate the preview display and preview loading components
