+++ apps/newsletters-ui/src/app/components/TemplateList.tsx
@@ -8,7 +8,6 @@ import {
 	Typography,
 } from '@mui/material';
 import type { RenderingTemplate } from '../loaders/rendering-templates';
-import { TemplatePreview } from './TemplatePreview';
 
 interface Props {
 	templates: RenderingTemplate[];
@@ -71,8 +70,6 @@ export const TemplateList = ({ templates }: Props) => {
 					<div>If in doubt, please contact the dev team.</div>
 				</Alert>
 			</Box>
-
-			<TemplatePreview />
 		</Container>
 	);
 };
