+++ apps/newsletters-ui/src/app/components/TemplatePreview.tsx
@@ -1,98 +1,19 @@
 import RefreshIcon from '@mui/icons-material/Refresh';
 import { Box, Button, ButtonGroup, Container } from '@mui/material';
-import { useCallback, useEffect, useState } from 'react';
-import type { NewsletterData } from '@newsletters-nx/newsletters-data-client';
-import { fetchApiData, fetchPostApiData } from '../api-requests/fetch-api-data';
+import { useState } from 'react';
 
 interface Props {
-	identityName?: string;
-	newsletterData?: NewsletterData;
+	html?: string;
 	minHeight?: number;
+	isLoading?: boolean;
 }
 
 export const TemplatePreview = ({
-	identityName,
-	newsletterData,
+	html,
 	minHeight = 800,
+	isLoading,
 }: Props) => {
 	const [frameWidth, setFrameWidth] = useState(360);
-	const [content, setContent] = useState<string | undefined>(undefined);
-	const [fetchInProgress, setFetchInProgress] = useState(false);
-	const [madeInitialFetch, setMadeInitalFetch] = useState(false);
-	const [gotInitalContent, setGotInitialContent] = useState(false);
-	const [hasChangedSinceLastRequest, setHasChangeSinceLastRequest] =
-		useState(false);
-
-	const fetchData = useCallback(
-		async (isInitial: boolean) => {
-			setFetchInProgress(true);
-			if (!identityName && !newsletterData) {
-				setFetchInProgress(false);
-				setContent(undefined);
-				return;
-			}
-
-			if (identityName) {
-				const data = await fetchApiData<{ content: string }>(
-					`/api/rendering-templates/preview/${identityName}`,
-				);
-				setFetchInProgress(false);
-				if (isInitial) {
-					setGotInitialContent(true);
-				}
-				if (data) {
-					setContent(data.content);
-				}
-			}
-
-			if (newsletterData) {
-				const data = await fetchPostApiData<{ content: string }>(
-					`/api/rendering-templates/preview`,
-					newsletterData,
-				);
-				setFetchInProgress(false);
-				if (isInitial) {
-					setGotInitialContent(true);
-				}
-				if (data) {
-					setContent(data.content);
-				}
-			}
-		},
-		[identityName, newsletterData],
-	);
-
-	// fetch on initial render
-	useEffect(() => {
-		if (madeInitialFetch) {
-			return;
-		}
-		setMadeInitalFetch(true);
-		void fetchData(true);
-	}, [fetchData, madeInitialFetch]);
-
-	// set the flag to mark changes to the data
-	useEffect(() => {
-		if (!gotInitalContent) {
-			return;
-		}
-		setHasChangeSinceLastRequest(true);
-	}, [newsletterData, gotInitalContent]);
-
-	// every five seconds, fetch data if there have been any changes
-	// since the last fetch
-	useEffect(() => {
-		const timer = setInterval(() => {
-			if (hasChangedSinceLastRequest) {
-				setHasChangeSinceLastRequest(false);
-				void fetchData(false);
-			}
-		}, 5000);
-
-		return () => {
-			clearInterval(timer);
-		};
-	}, [hasChangedSinceLastRequest, fetchData]);
 
 	const WidthButton = (props: { width: number }) => (
 		<Button
@@ -121,12 +42,9 @@ export const TemplatePreview = ({
 				<WidthButton width={650} />
 			</ButtonGroup>
 
-			{(fetchInProgress || hasChangedSinceLastRequest) && (
+			{isLoading && (
 				<Box sx={{ position: 'absolute', top: 0, right: 0 }}>
-					<RefreshIcon
-						sx={{ fontSize: 160 }}
-						color={fetchInProgress ? 'success' : 'warning'}
-					/>
+					<RefreshIcon sx={{ fontSize: 160 }} />
 				</Box>
 			)}
 
@@ -137,14 +55,14 @@ export const TemplatePreview = ({
 				justifyContent={'center'}
 				paddingY={2}
 			>
-				{content && (
+				{html && (
 					<iframe
 						style={{
 							width: frameWidth,
 							border: 'none',
 						}}
 						title="preview"
-						srcDoc={content}
+						srcDoc={html}
 					/>
 				)}
 			</Box>
